Moved message and mux directly into ws
