+++ ws/mux.go
@@ -1,26 +1,22 @@
-package mux
-
-import (
-	"github.com/NilsG-S/antifreeze-back-end/common"
-)
+package ws
 
 // maps allow concurrent access so this is fine
 type Mux struct {
-	routes map[string]func(common.Message, interface{})
+	routes map[string]func(Message, interface{})
 }
 
 func New() *Mux {
 	return &Mux{
-		routes: make(map[string]func(common.Message, interface{})),
+		routes: make(map[string]func(Message, interface{})),
 	}
 }
 
-func (m *Mux) AddRoute(route string, handler func(common.Message, interface{})) {
+func (m *Mux) AddRoute(route string, handler func(Message, interface{})) {
 	m.routes[route] = handler
 }
 
 // TODO: this needs to handle operation as well as route. combine the two?
-func (m *Mux) Handle(mes common.Message, context interface{}) {
+func (m *Mux) Handle(mes Message, context interface{}) {
 	handler, ok := m.routes[mes.GetSub()]
 
 	if !ok {
